==> notebooks/01_DataScraping.ipynb <==
# DataScraping

In this notebook I scrape the necessary data for these studies from onequestionshootout.xyz

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import time

# Delay used to be kind to the server owner and keep the requests at a human level
TIME_DELAY = 20
NUMBER_OF_SERIES = 16

def scrape_episodes(series_limit = 15):
    """
    Will scrape the html table for series <SERIES> from http://onequestionshootout.xyz/episodes/series_<SERIES>.htm 
    Function iterates from 1,2,...,series_limit and stacks each dataframe
    Additionally will scrape celebrity special episodes from http://onequestionshootout.xyz/episodes/celebrity_specials.htm
    Returns the stacked dataframe
    """

    df_list = []
    
    for series in range(1,series_limit+1):
        
        print("Scraping series: {}".format(series))
        if series > 1:
            time.sleep(TIME_DELAY)
                        
        data_url = "http://onequestionshootout.xyz/episodes/series_{}.htm".format(series)
        try:
            tables = pd.read_html(data_url)
        except URLError:
            break
        
        df_series = tables[1]
        df_series["Series"] = series
        df_series["isCelebrity"] = False
        
        df_list.append(df_series)
        
    print("Scraping series: Celebrity Specials")        
    time.sleep(TIME_DELAY)
    data_url = "http://onequestionshootout.xyz/episodes/celebrity_specials.htm"
    try:
        tables = pd.read_html(data_url)
        df_series = tables[1]
        df_series["Series"] = 0 # 0 used for celebrities to keep column numerical
        df_series["isCelebrity"] = True
        df_list.append(df_series)
    except URLError:
        pass
        
    df_episodes = pd.concat(df_list)
    print("\nTotal Episodes Scraped: {}".format(len(df_episodes)))
    return df_episodes

df_episodes = scrape_episodes(NUMBER_OF_SERIES)

df_episodes.to_csv("../data/all_episodes.csv",index=False)

def scrape_players(series_limit = 15):
    """
    Will scrape the html table for players <SERIES> from http://onequestionshootout.xyz/players/series_<SERIES>.htm 
    Function iterates from 1,2,...,series_limit and stacks each dataframe
    Additionally will scrape celebrity special players from http://onequestionshootout.xyz/players/celebrity_specials.htm
    Returns the stacked dataframe
    """

    df_list = []
    
    for series in range(1,series_limit+1):
        
        print("Scraping series: {}".format(series))
        if series > 1:
            time.sleep(TIME_DELAY)
                        
        data_url = "http://onequestionshootout.xyz/players/series_{}.htm".format(series)
        try:
            tables = pd.read_html(data_url)
        except URLError:
            break
        
        df_series = tables[1]
        df_series["Series"] = series
        df_series["isCelebrity"] = False
        
        df_list.append(df_series)

    print("Scraping series: Celebrity Specials")        
    time.sleep(TIME_DELAY)
    data_url = "http://onequestionshootout.xyz/players/celebrity_specials.htm"
    try:
        tables = pd.read_html(data_url)
        df_series = tables[1]
        df_series["Series"] = 0 # 0 used for celebrities to keep column numerical
        df_series["isCelebrity"] = True
        df_list.append(df_series)
    except URLError:
        pass        
                        
    df_players = pd.concat(df_list)
    print("\nTotal Players Scraped: {}".format(len(df_players)))
    return df_players

df_players = scrape_players(NUMBER_OF_SERIES)

# Create empty columns for Gender, Age, Occupation to be filled in manually later...
#empty_columns = pd.DataFrame('', index=df_players.index, columns=['Gender', 'Age', 'Occupation'])

# Insert the empty columns between 'Name' and 'CashBuilder'
#df_players = pd.concat([df_players.iloc[:, :3], empty_columns, df_players.iloc[:, 3:]], axis=1)

df_players.to_csv("../data/all_players.csv",index=False)
