==> notebooks/02_DataCleaning_Episodes.ipynb <==
# Data Cleaning - Episodes Data

This notebook will perform some data cleaning on the episodes dataframe. Changes mostly involve parsing the raw strings into a useful form for EDA and Modelling studies.

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt

Quick check that we havent lossed/duplicated data somewhere, check that the number of players is 4x the number of episodes.

df_episodes = pd.read_csv("../data/all_episodes.csv")
df_players = pd.read_csv("../data/all_players.csv")
# sanity check episodes x 4 = players
print("Episodes: {}".format(len(df_episodes)))
print("Players: {}".format(len(df_players)))
print("Length Check: {}".format(len(df_episodes)*4 == len(df_players)))

## Dataframe Inspection

df_episodes.head()

df_episodes.tail()

## Drop Columns

For these studies we don't need to track information on Chaser performance, will drop columns relating to this. We also drop the "FinalChaseVideo" and "Series &Episode" columns ofr being useless to us.

df_episodes = df_episodes.drop(columns=['Push-backAttempts', 'Push-backsTaken', 'ChaserAccuracy', 'ChaserSpeed', 'FinalChaseVideo', 'Series &Episode'])

## Rename Columns

Rename some columns to be more easily used in code

df_episodes = df_episodes.rename(columns={'No. OfPlayers InFinal Chase':'NPlayersFinal'})

df_episodes.columns

## Check All Unique

Simply making sure there are no duplicate episodes. Using Date, Episode and Series as the ID

df_episodes[df_episodes[["Date","Episode","Series"]].duplicated(keep=False)]

## Check Player Numbers

Check how many players are listed in the "Team" column for each episode. Expect for them all to be 4.

Should be a little careful of celebrity specials often have the theme listed first:
THE CHASE FOR SOCCER AID: Dion Dublin, Lianne Sanderson, David James, Shaun Williamson

To count names, first split on ":", then split on ","

df_episodes["Team"].apply(lambda x: len([name for name in x.split(":")[-1].split(",") if len(name) > 0])).value_counts()

## Chaser Episode Count

Check how many episodes each chaser has featured in

# Check for all chasers encountered
df_episodes["Chaser"].value_counts()

Seems this single entry "Anne Hegerty *" occured for the 1000th episode special where 5 chasers were used, only one for the final chase. For these studies it is probably best to remove this datapoint.

df_episodes = df_episodes[~((df_episodes["Series"] == 10) & (df_episodes["Episode"] == 54))]

## Target Score

Check target scores make sense.

Often we see "+1" is being appended when the Chaser fails to get their first question and the team is able to push back by one. For this analysis I will only reference data from before the chasers final chase, hence they should be dropped.

# Check for all targets
df_episodes["Target"].value_counts()

df_episodes["Target"] = df_episodes["Target"].apply(lambda x: int(str(x).split("+")[0]))

## PrizeFund string to Float

Want to Convert the string "£20,200" into numeric "20200.00"

df_episodes["PrizeFund"] = df_episodes["PrizeFund"].str.replace("£","").str.replace(",","").astype('float')

df_episodes["PrizeFund"].head()

## Parse Chaser Win Variables

The format of the "Winner" column is:
- Chaser by 0:07
- Team by 6

Want to extract this data:
- ChaserWin (Bool)
- ChaserWinBy (int: seconds), 0 for Chaser Loss
- ChaserLoseBy (int: questions), 0 for Chaser Win

df_episodes["ChaserWin"] = df_episodes["Winner"].str.contains("Chaser")

# Cases when chaser wins, how many seconds left on clock?
df_episodes["ChaserWinBy"] = df_episodes["Winner"].apply(lambda x: x.split("Chaser by ")[-1])
df_episodes.loc[df_episodes["ChaserWinBy"].str.contains("Team"),"ChaserWinBy"] = "0:00"
# turns mm:ss into int seconds
df_episodes["ChaserWinBy"] = df_episodes["ChaserWinBy"].apply(lambda x: pd.to_timedelta("00:" + x).total_seconds()).astype(int)

# Cases when chaser loses, how many questions did they have left?
df_episodes["ChaserLoseBy"] = df_episodes["Winner"].apply(lambda x: x.split("Team by ")[-1])
df_episodes.loc[df_episodes["ChaserLoseBy"].str.contains("Chaser"),"ChaserLoseBy"] = "0"


df_episodes[["Winner", "ChaserWin", "ChaserWinBy", "ChaserLoseBy"]].head(20)

Finally drop Winner string

df_episodes = df_episodes.drop(columns=['Winner'])

## Clean Team String

Some of the specials have labels for what the episode is themed on, will remove it for easier processing 

THE CHASE FOR SOCCER AID: Dion Dublin, Lianne Sanderson, David James, Shaun Williamson

to

Dion Dublin, Lianne Sanderson, David James, Shaun Williamson

Performed by splitting on ":" and keeping the last element, then using lstrip to remove leading spaces.

df_episodes["Team"] = df_episodes["Team"].apply(lambda x: x.split(":")[-1].lstrip())

df_episodes["Team"].tail()

# Final Inspection and Save

df_episodes.head()

df_episodes.tail()

df_episodes.to_csv('../data/cleaned_episodes.csv', index=False)
