==> notebooks/02_DataCleaning_Players.ipynb <==
# Data Cleaning - Player Data

This notebook will perform some data cleaning on the episodes dataframe. Changes mostly involve parsing the raw strings into a useful form for EDA and Modelling studies.

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt

Quick check that we havent lossed/duplicated data somewhere, check that the number of players is 4x the number of episodes.

df_episodes = pd.read_csv("../data/all_episodes.csv")
df_players = pd.read_csv("../data/all_players.csv")
# sanity check episodes x 4 = players
print("Episodes: {}".format(len(df_episodes)))
print("Players: {}".format(len(df_players)))
print("Length Check: {}".format(len(df_episodes)*4 == len(df_players)))

## Dataframe Inspection

df_players.head()

df_players.tail()

## Drop Columns

For these studies we don't need to track information on the players contribution to the 

df_players = df_players.drop(columns=['Chaser', 'FC CorrectAnswers', 'FC Winner', 'Amount WonBy Player', 'bgc', 'isCelebrity'])

df_players.head()

## PrizeFund string to Float

When a player answers no questions in the cash builder the Chaser can offer no money for the lower offer, marked as "no offer". There are also some columns marked as "unknown". In rarer cases we have no HigherOffer also with just a blank "£" given.

These shall be replaced with np.nan as XGBoost should be able to handle missing values like this

df_players.loc[df_players["LowerOffer"] == "no offer", "LowerOffer"] = np.nan
df_players.loc[df_players["LowerOffer"] == "unknown", "LowerOffer"] = np.nan
df_players.loc[df_players["HigherOffer"] == "£", "HigherOffer"] = np.nan

Want to Convert the string "£20,200" into numeric "20200.00" and also remove the selected offer notation on the ChosenOffer column, i.e. " /\" or " =". We can infer this by column comparisons later in the analysis. 

def strip_offer_string(offer_string):
    try:
        return float(offer_string)
    except ValueError:
        # keep only first split of string
        output_string = offer_string.split()[0]

        # remove currency symbol and commas
        output_string = output_string.replace("£", "").replace(",", "")
        # format pence to be 1/100th of a £...
        output_string = output_string.replace("p", "e-2")

        # if nothing survived, will just return zero
        if output_string == "":
            return 0

        return float(output_string)



df_players["CashBuilder"] = df_players["CashBuilder"].apply(lambda x: strip_offer_string(x))
df_players["LowerOffer"] = df_players["LowerOffer"].apply(lambda x: strip_offer_string(x))
df_players["HigherOffer"] = df_players["HigherOffer"].apply(lambda x: strip_offer_string(x))
df_players["ChosenOffer"] = df_players["ChosenOffer"].apply(lambda x: strip_offer_string(x))

We also want to store information of whether a given offer was chosen by the player:
- LowerOffer_Selected (boolean): Did the player select the lower offer?
- MiddleOffer_Selected (boolean): Did the player select the middle offer?
- HigherOffer_Selected (boolean): Did the player select the higher offer?

df_players["LowerOffer_Selected"] = df_players["ChosenOffer"] == df_players["LowerOffer"]
df_players["MiddleOffer_Selected"] = df_players["ChosenOffer"] == df_players["CashBuilder"]
df_players["HigherOffer_Selected"] = df_players["ChosenOffer"] == df_players["HigherOffer"]

## Format HeadToHead Result

We want to format the "HTHResult" column into two separate columns:
- isHome (boolean): Was the player successfull in getting back home?  
- LadderDifference (int): At the end of the HTH what was the difference between the chaser and players ladder position? Positive for Home, Negative for Caught.  

df_players["isHome"] = df_players["HTHResult"].str.contains("Home")

def strip_hth_string(offer_string):
    
    # keep only last split of string
    output_string = offer_string.split()[-1]
    
    if output_string == "Home":
        return 1
    if output_string == "Caught":
        return -1
    
    output_string = output_string.replace("+","")

    if output_string == "":
        return 1
    
    return int(output_string)

df_players["LadderDifference"] = df_players["HTHResult"].apply(lambda x: strip_hth_string(x))

df_players["LadderDifference"].value_counts()

Remove the HTHResult column

df_players = df_players.drop(columns=['HTHResult'])

# Final Inspection and Save

df_players.head()

df_players.tail()

df_players.to_csv('../data/cleaned_players.csv', index=False)
